Use _app_ctx_stack and _request_ctx_stack top as context
This is according to a recommendation on extension development in Flask
documentation. This also fixes an issue where force_locale context leaks
between two threads which are running in app context rather than request
context. `current_app` is a proxy to the application object, while the
object on the relevant stack is pushed when application context is
created.